[UHYU-102] refactor: guest 도메인 분리
