+++ src/main/java/com/ureca/uhyu/domain/mymap/service/MyMapService.java
@@ -105,17 +105,6 @@ public class MyMapService {
         return MyMapRes.from(myMapList, storeList, isMine);
     }
 
-    public MyMapRes findMyMapByUUIDWithGuest(String uuid) {
-        MyMapList myMapList = myMapListRepository.findByUuid(uuid).orElseThrow(() -> new GlobalException(ResultCode.MY_MAP_LIST_NOT_FOUND));
-        List<MyMap> myMaps = myMapRepository.findByMyMapList(myMapList);
-
-        List<MapRes> storeList = myMaps.stream()
-                .map(myMap -> MapRes.from(myMap.getStore()))
-                .toList();
-
-        return MyMapRes.from(myMapList, storeList, false);
-    }
-
     public BookmarkedMyMapRes findMyMapListWithIsBookmarked(User user, Long storeId) {
         Store store = storeRepository.findById(storeId)
                 .orElseThrow(() -> new GlobalException(ResultCode.NOT_FOUND_STORE));
