+++ src/main/java/com/ureca/uhyu/domain/recommendation/service/RecommendationService.java
@@ -1,7 +1,5 @@
 package com.ureca.uhyu.domain.recommendation.service;
 
-import com.ureca.uhyu.domain.brand.entity.Brand;
-import com.ureca.uhyu.domain.recommendation.dto.response.GuestRecommendationRes;
 import com.ureca.uhyu.domain.recommendation.dto.response.RecommendationRes;
 import com.ureca.uhyu.domain.recommendation.repository.RecommendationRepository;
 import com.ureca.uhyu.domain.user.entity.User;
@@ -34,21 +32,6 @@ public class RecommendationService {
                             r.getRank()
                     );
                 })
-                .toList();
-    }
-
-    public List<GuestRecommendationRes> getTop3PopularBrandsForGuest() {
-        List<Brand> brands = recommendationRepository.findTop3BrandByVisitCountFromHistory();
-
-        if (brands == null || brands.isEmpty()) {
-            throw new GlobalException((ResultCode.BRAND_ID_IS_NULL));
-        }
-        return brands.stream()
-                .map(brand -> GuestRecommendationRes.from(
-                        brand.getId(),
-                        brand.getBrandName(),
-                        brand.getLogoImage()
-                ))
                 .toList();
     }
 }
